feat: ArticleDatabase 객체를 Spring Bean으로 등록 - static class로 선언된 클래스를 spring bean으로 등록
