+++ src/main/java/codesquad/springcafe/main/controller/MainController.java
@@ -2,7 +2,7 @@ package codesquad.springcafe.main.controller;
 
 
 import codesquad.springcafe.articles.service.ArticleService;
-import model.article.Article;
+import codesquad.springcafe.model.article.Article;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
