+++ src/main/java/codesquad/springcafe/model/user/dto/UserCredentialDto.java
@@ -1,4 +1,4 @@
-package model.user.dto;
+package codesquad.springcafe.model.user.dto;
 
 public class UserCredentialDto {
     private final String password;
