+++ src/main/java/codesquad/springcafe/model/user/dto/UserPreviewDto.java
@@ -1,4 +1,4 @@
-package model.user.dto;
+package codesquad.springcafe.model.user.dto;
 
 public class UserPreviewDto {
     private final String userId;
