+++ src/main/java/codesquad/springcafe/model/user/dto/UserUpdateData.java
@@ -1,4 +1,4 @@
-package model.user.dto;
+package codesquad.springcafe.model.user.dto;
 
 public class UserUpdateData {
     private final String newName;
